requirements.txt
# Core API requirements
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# HTTP and web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# GitHub integration
PyGithub>=2.1.0

# PDF generation
# Note: pdflatex must be installed separately via texlive or similar

# Type hints and async support
typing-extensions>=4.7.0
python-multipart>=0.0.6

# API documentation (included with FastAPI but specified for version control)
starlette>=0.27.0
openapi-schema-pydantic>=1.2.4

# Error handling and logging
python-json-logger>=2.0.7

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0